+++ tools/sim/bridge/metadrive/metadrive_process.py
@@ -2,17 +2,45 @@ import math
 import numpy as np
 
 from collections import namedtuple
-
 from multiprocessing.connection import Connection
+
+from metadrive.engine.core.engine_core import EngineCore
+from metadrive.engine.core.image_buffer import ImageBuffer
 from metadrive.envs.metadrive_env import MetaDriveEnv
+from metadrive.obs.image_obs import ImageObservation
 
 from openpilot.common.realtime import Ratekeeper
 from openpilot.tools.sim.lib.common import vec3
 
 metadrive_state = namedtuple("metadrive_state", ["velocity", "position", "bearing", "steering_angle"])
 
+def apply_metadrive_patches():
+  # By default, metadrive won't try to use cuda images unless it's used as a sensor for vehicles, so patch that in
+  def add_image_sensor_patched(self, name: str, cls, args):
+    if self.global_config["image_on_cuda"]:# and name == self.global_config["vehicle_config"]["image_source"]:
+        sensor = cls(*args, self, cuda=True)
+    else:
+        sensor = cls(*args, self, cuda=False)
+    assert isinstance(sensor, ImageBuffer), "This API is for adding image sensor"
+    self.sensors[name] = sensor
+
+  EngineCore.add_image_sensor = add_image_sensor_patched
+
+  # we aren't going to use the built-in observation stack, so disable it to save time
+  def observe_patched(self, vehicle):
+    return self.state
+
+  ImageObservation.observe = observe_patched
+
+  def arrive_destination_patch(self, vehicle):
+    return False
+
+  MetaDriveEnv._is_arrive_destination = arrive_destination_patch
+
 
 def metadrive_process(dual_camera: bool, config: dict, camera_send: Connection, controls_recv: Connection, state_send: Connection):
+  apply_metadrive_patches()
+
   env = MetaDriveEnv(config)
   env.reset()
 
@@ -39,14 +67,20 @@ def metadrive_process(dual_camera: bool, config: dict, camera_send: Connection, 
     state_send.send(state)
 
     while controls_recv.poll(0):
-      steer_angle, gas = controls_recv.recv()
+      steer_angle, gas, reset = controls_recv.recv()
 
       steer_metadrive = steer_angle * 1 / (env.vehicle.MAX_STEERING * steer_ratio)
       steer_metadrive = np.clip(steer_metadrive, -1, 1)
 
       vc = [steer_metadrive, gas]
 
-    env.step(vc)
+      if reset:
+        env.reset()
+
+    obs, _, terminated, _, info = env.step(vc)
+
+    if terminated:
+      env.reset()
 
     #if dual_camera:
     #  wide_road_image = get_cam_as_rgb("rgb_wide")
